% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barchart.R
\name{barchart}
\alias{barchart}
\title{Create a bar chart with labels}
\usage{
barchart(df, x, y, label, x_label, y_label, title, vjust = 0.5, hjust = 0.5)
}
\arguments{
\item{df}{A data frame with the observations}

\item{x}{The name of the x-axis}

\item{y}{The name of the y-axis}

\item{label}{The labels to display}

\item{x_label}{The label of the x-axis}

\item{y_label}{The label of the y-axis}

\item{title}{The title of the plot}

\item{vjust}{To adjust vertically the labels}

\item{hjust}{To adjust horizontally the labels}
}
\description{
This function allows you to create a bar chart and display the labels
}
\examples{
library(ggplot2)
df_gg <- count_percentage(iris, Species)
barchart(df = df_gg, x = Species, y = percentage, label = percentage,
x_label = "Species", y_label = "Percentage",
title = "Distribution of Species in Iris dataset", vjust = -0.25)
}
\keyword{label}
\keyword{plot,}
