% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_se.R
\name{compute_se}
\alias{compute_se}
\title{Compute the spherical equivalent}
\usage{
compute_se(sph = 0, cyl = 0, digits = 3)
}
\arguments{
\item{sph}{The sphere}

\item{cyl}{The cylinder}

\item{digits}{The number of decimals needed}
}
\description{
This function allows you to compute the spherical equivalent of
a refraction.
}
\examples{
compute_se(-1, -0.5)
}
\keyword{compute,}
\keyword{spherical}
