% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_percentage.R
\name{count_percentage}
\alias{count_percentage}
\title{Count and compute percentage}
\usage{
count_percentage(df, col_name, digits = 1)
}
\arguments{
\item{df}{A data frame with the observations}

\item{col_name}{The name of the column containing the observations}

\item{digits}{The number of decimals needed for the percentage}
}
\description{
This function allows you to count the observations per group
and compute the associated percentage.
}
\examples{
count_percentage(iris, Species)
}
\keyword{count,}
\keyword{percentage}
