% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_percentage_multiple.R
\name{count_percentage_multiple}
\alias{count_percentage_multiple}
\title{Count and compute percentage on multiple columns}
\usage{
count_percentage_multiple(df, ..., digits = 1)
}
\arguments{
\item{df}{A data frame with the observations}

\item{...}{The name of the columns containing the observations}

\item{digits}{The number of decimals needed for the percentage}
}
\description{
This function allows you to count the observations per group on multiple columns
and compute the associated percentage.
}
\examples{
count_percentage_multiple(iris, Species)
}
\keyword{count,}
\keyword{percentage}
