% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_charts_ppt.R
\name{export_charts_ppt}
\alias{export_charts_ppt}
\title{Export a list of charts made in ggplot to Powerpoint}
\usage{
export_charts_ppt(
  listP,
  export_path,
  template_path = NULL,
  layout = "Title and Content",
  master = "Office Theme"
)
}
\arguments{
\item{listP}{a list with the ggplot charts}

\item{export_path}{The path where to export the Powerpoint}

\item{template_path}{The path of the Powerpoint template}

\item{layout}{Name of the Powerpoint layout}

\item{master}{Name of the Powerpoint master}
}
\description{
This function allows you export a list of charts made in ggplot to Powerpoint.
}
\examples{
listP = list()
listP[["first_chart"]] <- ggplot(data = iris ) +
geom_point(mapping = aes(Sepal.Length, Petal.Length))
export_charts_ppt(listP, export_path = "test.pptx")
}
\keyword{chart,}
\keyword{ppt}
